bug: Remove unnecessary send experience from agent
Remove unnecessary send experience from agent method act(), since if we reached timestep limit, then reset is perform by the client. It is sufficient to send experience only in reward_ and_reset() method and check the timestep limit to be sure that the trajectory gets terminal by limit (as natural end). Additional metric (server latency) is also removed to reduce the I/O rate and increade performance for the experiment.